Created but didn`t checked database ScannedGroup insert. Need to check. Replaced int id to Long id.

+++ app/src/main/java/com/crystal2033/qrextractor/scanner_feature/data/remote/api/KeyboardApi.kt
@@ -8,5 +8,5 @@ import retrofit2.http.Path
 interface KeyboardApi {
 
     @GET(value = "api/keyboards/{id}")
-    suspend fun getKeyboard(@Path(value = "id") id: Int) : Response<KeyboardDto?>
+    suspend fun getKeyboard(@Path(value = "id") id: Long) : Response<KeyboardDto?>
 }
