+++ app/src/main/java/com/crystal2033/qrextractor/scanner_feature/data/util/ScannedTableNameAndId.kt
@@ -1,4 +1,4 @@
 package com.crystal2033.qrextractor.scanner_feature.data.util
 
 
-data class ScannedTableNameAndId(val tableName: String, val id: Int)
+data class ScannedTableNameAndId(val tableName: String, val id: Long)
