+++ app/src/main/java/com/crystal2033/qrextractor/scanner_feature/domain/model/QRScannableData.kt
@@ -7,5 +7,5 @@ package com.crystal2033.qrextractor.scanner_feature.domain.model
  */
 interface QRScannableData {
     fun getDatabaseTableName():DatabaseObjectTypes
-    fun getDatabaseID(): Int
+    fun getDatabaseID(): Long
 }
