+++ app/src/main/java/com/crystal2033/qrextractor/scanner_feature/domain/repository/PersonRepository.kt
@@ -5,5 +5,5 @@ import com.crystal2033.qrextractor.scanner_feature.domain.model.Person
 import kotlinx.coroutines.flow.Flow
 
 interface PersonRepository {
-    fun getPerson(id: Int): Flow<Resource<Person>>
+    fun getPerson(id: Long): Flow<Resource<Person>>
 }
