+++ app/src/main/java/com/crystal2033/qrextractor/scanner_feature/domain/use_case/concrete_use_case/GetKeyboardFromQRCodeUseCase.kt
@@ -11,7 +11,7 @@ import kotlinx.coroutines.flow.flow
 class GetKeyboardFromQRCodeUseCase(
     private val repository: KeyboardRepository
 ) : GetDataFromQRCodeUseCase {
-    override operator fun invoke(id: Int): Flow<Resource<QRScannableData>> {
+    override operator fun invoke(id: Long): Flow<Resource<QRScannableData>> {
         if (id < 1) {
             return flow { }
         }
